% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_count.R
\name{after_count}
\alias{after_count}
\alias{after_inner_count}
\alias{after_left_count}
\alias{after_right_count}
\alias{after_full_count}
\title{Join tables to get the count of events following another}
\usage{
after_count(x, y, by_time, by_user, mode = "inner", type = "first",
  max_gap = NULL, min_gap = NULL, gap_col = FALSE, suffix = c(".x",
  ".y"))

after_inner_count(x, y, by_time, by_user, type, max_gap = NULL,
  min_gap = NULL, gap_col = FALSE, suffix = c(".x", ".y"))

after_left_count(x, y, by_time, by_user, type, max_gap = NULL,
  min_gap = NULL, gap_col = FALSE, suffix = c(".x", ".y"))

after_right_count(x, y, by_time, by_user, type, max_gap = NULL,
  min_gap = NULL, gap_col = FALSE, suffix = c(".x", ".y"))

after_full_count(x, y, by_time, by_user, type, max_gap = NULL,
  min_gap = NULL, gap_col = FALSE, suffix = c(".x", ".y"))
}
\arguments{
\item{x}{A tbl representing the first event to occur in the funnel.}

\item{y}{A tbl representing an event to occur in the funnel.}

\item{by_time}{A character vector to specify the time columns in x and y.
This would typically be a datetime or a date column. These columns are used to
filter for time y being after time x.}

\item{by_user}{A character vector to specify the user or identity columns in
x and y.}

\item{mode}{The method used to join: "inner", "full", "anti", "semi",
"right", "left". Each also has its own function, such as
\code{after_inner_join}.}

\item{type}{The type of event you want from x - can be "first", "last", or "any."}

\item{max_gap}{Optional: the maximum gap allowed between events. Can be a
integer representing the number of seconds or a difftime object, such as
\code{as.difftime(2, units = "hours")}.}

\item{min_gap}{Optional: the maximum gap allowed between events. Can be a
integer representing the number of seconds or a difftime object, such as
\code{as.difftime(2, units = "hours")}.}

\item{gap_col}{Whether to include a numeric column, \code{.gap},
with the time difference in seconds between the events.}

\item{suffix}{If there are non-joined duplicate variables in x and y,
these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}
}
\value{
a tbl with the user column, timestamp from the first table, and number of events from the second tbl following that timestamp
}
\description{
Join tables to get the count of events following another
}
