% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after_join_all.R
\name{after_join_all}
\alias{after_join_all}
\title{View result for each type of afterjoin}
\usage{
after_join_all(x, y, by_user, by_time, mode = "inner", ...)
}
\arguments{
\item{x}{A tbl representing the first event to occur in the funnel.}

\item{y}{A tbl representing an event to occur in the funnel.}

\item{by_user}{A character vector to specify the user or identity columns in
x and y.}

\item{by_time}{A character vector to specify the time columns in x and y.
This would typically be a datetime or a date column.
These columns are used to filter for time y being after time x.}

\item{mode}{The method used to join: "inner", "full", "anti", "semi",
"right", "left".}

\item{...}{any additional arguments}
}
\description{
View result for each type of afterjoin
}
