% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_seconds.R
\name{as_seconds}
\alias{as_seconds}
\title{Title}
\usage{
as_seconds(x, sql = FALSE)
}
\arguments{
\item{x}{a difftime object}

\item{sql}{set to TRUE if you're working with remote tables and using dbplyr}
}
\value{
a difftime object in seconds
}
\description{
Title
}
