% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_events.R
\name{distinct_events}
\alias{distinct_events}
\title{Distinct events}
\usage{
distinct_events(.data, time_col, user_col, type)
}
\arguments{
\item{.data}{a dataset, either local or remote}

\item{time_col}{the name of the time column}

\item{user_col}{the name of the user identifying column}

\item{type}{the type of after_join ("first-first", "first-firstafter", etc.)}
}
\description{
Distinct events
}
