% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_funnel.R
\name{funnel_start}
\alias{funnel_start}
\title{Start a funnel}
\usage{
funnel_start(tbl, event_type, event, tstamp, user)
}
\arguments{
\item{tbl}{A table of different events and timestamps}

\item{event_type}{The first event in the funnel}

\item{event}{The name of the column with the event_type}

\item{tstamp}{The name of the column with the timestamps of the events}

\item{user}{The name of the column indicating the user who did the event}
}
\description{
Start a funnel
}
