% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_funnel.R
\name{funnel_step}
\alias{funnel_step}
\alias{funnel_steps}
\title{Continue to funnel}
\usage{
funnel_step(tbl, moment_type, type, name = moment_type, optional = FALSE, ...)

funnel_steps(tbl, moment_types, type, ...)
}
\arguments{
\item{tbl}{A table of different moments and timestamps}

\item{moment_type}{The next moment in the funnel}

\item{type}{The type of after_join (e.g. "first-first", "any-any")}

\item{name}{If you want a custom name instead of the moment_type; needed if the moment type is already in the sequence}

\item{optional}{Whether this step in the funnel should be optional. If so, the following step will also try joining
in a way that skips this step. Note that multiple optional steps in a row aren't supported.}

\item{...}{Extra arguments passed on to \link{after_left_join}. For \code{funnel_steps}, these are passed on to \code{funnel_step}.}

\item{moment_types}{For \code{funnel_steps}, a character vector of
moment types, which are applied in order}
}
\description{
Continue to funnel
}
