% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_funnel.R
\name{funnel_steps}
\alias{funnel_steps}
\title{Multiple}
\usage{
funnel_steps(tbl_funnel, moment_types, type, ...)
}
\arguments{
\item{steps}{For \code{funnel_steps}, a vector}
}
\description{
Multiple
}
