% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landed}
\alias{landed}
\title{Example dataset of landing events}
\format{
A data frame with 9 rows and 2 variables:
\describe{
  \item{user_id}{A numeric column for identifying people}
  \item{timestamp}{A date column for the date the landing happened}
}
}
\usage{
landed
}
\description{
An example dataset for trying out after_join. The variables are as follows:
}
\keyword{datasets}
