% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-funnel-reclass.R
\name{reclass}
\alias{reclass}
\title{Copy class and attributes from the original version of an object to a modified version.}
\usage{
reclass(x, result)
}
\arguments{
\item{x}{The original object, which has a class/attributes to copy}

\item{result}{The modified object, which is / might be missing the
class/attributes.}
}
\value{
\code{result}, now with class/attributes restored.
}
\description{
Copied over from https://github.com/tidyverse/dplyr/issues/719
}
