% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_funnel.R
\name{summarize_conversions}
\alias{summarize_conversions}
\title{Summarize Left-joined table into conversion count}
\usage{
summarize_conversions(x, time_col_y = timestamp.y)
}
\arguments{
\item{x}{a data.frame with one row per user}

\item{time_col_y}{the name of the second event time column}
}
\value{
a table with three columns, `nb_starts`, `nb_conversions`, and `alternative.name`
}
\description{
Summarize Left-joined table into conversion count
}
