% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_funnel.R
\name{summarize_funnel}
\alias{summarize_funnel}
\title{Summarize after funnel start and funnel step(s)}
\usage{
summarize_funnel(tbl_funnel)
}
\arguments{
\item{tbl_funnel}{a table from funnel start and funnel step(s)}
}
\description{
Summarize after funnel start and funnel step(s)
}
