% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_funnel.R
\name{summarize_prop_tests}
\alias{summarize_prop_tests}
\title{Summarise after join funnel with proportion test}
\usage{
summarize_prop_tests(x, ..., ungroup = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{ungroup}{whether the table needs to be ungrouped}
}
\value{
a data.frame with proportion test results
}
\description{
Summarise after join funnel with proportion test
}
