% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_funnel.R
\name{funnel_start}
\alias{funnel_start}
\title{Start a funnel}
\usage{
funnel_start(tbl, moment_type, moment, tstamp, user)
}
\arguments{
\item{tbl}{A table of different moments and timestamps}

\item{moment_type}{The first moment in the funnel}

\item{moment}{The name of the column with the moment_type}

\item{tstamp}{The name of the column with the timestamps of the moments}

\item{user}{The name of the column indicating the user who did the moment}
}
\description{
Start a funnel
}
