% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_funnel.R
\name{summarize_conversions}
\alias{summarize_conversions}
\title{Summarize Left-joined table into conversion count}
\usage{
summarize_conversions(x, converted)
}
\arguments{
\item{x}{A tbl with one row per user}

\item{converted}{The name of the column representing whether the user converted
(treated as FALSE if NA or FALSE, otherwise TRUE)}
}
\value{
A table with columns for your groups, along with `nb_starts` and `nb_conversions`
}
\description{
Summarize Left-joined table into conversion count
}
